+++ src/modules/credential/mod.rs
@@ -6,6 +6,7 @@ pub mod approle;
 pub mod cert;
 pub mod ferrogate;
 pub mod fido2;
+pub mod login_audit_store;
 pub mod oidc;
 pub mod saml;
 pub mod token;